Redirecting the User after Successfully Signing In
After a user successfully signs in, redirect them back to the previous screen (whatever that happens to be). For example, if a user attempts to view the "Create Course" screen before they've signed in, they'll be redirected to the "Sign In" screen. After the user has successfully signed in, redirect them to the "Create Course" screen.
Use useLocation hook